+++ test/audio_detail_batch_metadata_test.dart
@@ -1,5 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:nameless_audio/app/state/audio_provider.dart';
+import 'package:nameless_audio/core/media/audio_detail.dart';
+import 'package:nameless_audio/core/media/card_info_field.dart';
+import 'package:nameless_audio/features/library/application/dlsite_metadata_query.dart';
 
 AudioDetail _detail({
   String rjCode = 'RJ123456',